Split the core atoms and interpreter up
Proc is not a built-in (i.e., parse-able) atom, so it's going with the
interpreter. The package name interp is subject to change because it's
weird, but it's a good starting point.

Main goal here is to make the interpreter entirely optional and allow
more backends, including a bytecode compiler. The interpreter is fun to
play with, but it's not useful in the long term.

Change-Id: I737776196c4e09dd339bc73f65bf0bbd56ed318a