Add lambda functions
Lambdas are fairly simple -- they're merely the storied binding names
and bodies to be evaluated in a duplicated context. That is, at the
time the lambda is created, the context used to create the lambda is
stored. Once called, the creation context is duplicated into
a flattened context and overlaid on top of the calling context. This is
to ensure that the names known to the lambda at the time of its
creation are bound for the duration of the lambda's existence, but no
longer than necessary.

Lambda arguments are named in a vector, as opposed to the scheme-like
list that's used by let and let*. Those two will also likely change
later.

Context has been changed to add an Evaler interface. Both Proc and
Lambda implement this, removing the need for Context to specifically
recognize callable types beyond those that explicitly implement that
behavior.

A Dup() method has been added to *Cons and Vector to allow
deep-cloning of data structures. This is only necessary right now since
the Go implementation allows mutability. This could possibly be ignored
entirely and just document the safety of modifying anything you don't
own, but then ownership needs to be explained as well. Still thinking
about it, but Dup made sense in the short term.

Locks have been added to *Context methods to eventually allow things
like parallel evaluation of let bindings and similar.

Let will now evaluate its values in a forked context -- this is to stop
let from evaluating a (set ..) call and modifying a context
inappropriately.

Closes #2. Closes #3.

Change-Id: I7ea3a65551032f992ac2ceb6a203f6146ba754fa